orientai_claude/requirements.txt
streamlit>=1.32
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
gspread>=6.0
google-auth>=2.28
google-auth-oauthlib>=1.2
reportlab>=4.1
python-docx>=1.1
matplotlib>=3.8
numpy>=1.26
